refactor(map): split territory-style-constants into domain-axis files
Break the catch-all territory-style-constants.ts into four domain-axis
modules (label, territory-fill, highlight, loading-animation) so that
each concern lives in its own file. Replace hardcoded label hex colors
with design-tokens references and rebuild zoom interpolation expressions
through the new buildZoomInterpolation helper.
