+++ apps/frontend/src/components/map/highlight-style-constants.ts
@@ -1,0 +1,5 @@
+import { roleColors } from '@world-history-map/design-tokens';
+
+export const HIGHLIGHT_COLOR = roleColors.selected;
+export const HIGHLIGHT_FILL_OPACITY = 0.2;
+export const HIGHLIGHT_LINE_WIDTH = 3.5;
