+++ apps/frontend/src/components/map/loading-animation-constants.ts
@@ -1,0 +1,1 @@
+export const LOADING_SPIN_DURATION = '3s';
