+++ apps/frontend/src/components/map/territory-fill-constants.ts
@@ -1,0 +1,1 @@
+export const TERRITORY_FILL_OPACITY = 0.7;
